refactored header component
